requirements.txt
asgiref==3.11.1
black==26.3.0
bleach==6.3.0
click==8.3.1
colorama==0.4.6
Django==5.2.12
django-rest-framework==0.1.0
djangorestframework==3.16.1
mypy_extensions==1.1.0
packaging==26.0
pathspec==1.0.4
pillow==12.1.1
platformdirs==4.9.4
pytokens==0.4.1
sqlparse==0.5.5
tzdata==2025.3
webencodings==0.5.1
